+++ core/SCActuatorControlUnit.h
@@ -69,8 +69,8 @@ namespace driver {
 			bool setPosition(bool sync = true);
 
                         //-----------utility methdo for the restore operation---------
-			bool moveAt(const std::string &name,double value,size_t size);
-			bool setPower(const std::string &name,bool value,size_t size);
+			bool moveAt(const std::string &name,double value,uint32_t size);
+			bool setPower(const std::string &name,bool value,uint32_t size);
 
                         const uint32_t *axID;
